requirements.txt
gspread==3.6.0
requests==2.20.0
Telethon==1.28.5
Unidecode==1.2.0
pyTelegramBotAPI==4.22.1
google-api-python-client==1.10.0
google-auth==2.14.1